feat: implementation of request parameter validation.

+++ src/main/java/jp/nlaboratory/MyBatisWithThreeTierArchitecture/application/controller/UserController.java
@@ -1,11 +1,17 @@
 package jp.nlaboratory.MyBatisWithThreeTierArchitecture.application.controller;
 
 import java.time.LocalDateTime;
+import jp.nlaboratory.MyBatisWithThreeTierArchitecture.application.exception.InvalidParameterException;
+import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.dto.UserCreateRequest;
 import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.dto.UserRequest;
 import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.dto.UserResponse;
+import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.dto.UserUpdateRequest;
 import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.entity.User;
 import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.service.UserService;
+import jp.nlaboratory.MyBatisWithThreeTierArchitecture.domain.utils.Converter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,13 +31,17 @@ public class UserController {
   /**
    * Search user.
    *
-   * @param userRequest UserSearchRequest (id)
+   * @param request UserRequest (id)
    * @return UserResponse (id, name, email, delFlg)
    */
   @GetMapping(value = "/search")
-  public UserResponse search(@ModelAttribute UserRequest userRequest)
+  public UserResponse search(@ModelAttribute @Validated UserRequest request, BindingResult result)
       throws Exception {
-    User user = userService.getUser(userRequest.getId());
+    if (result.hasErrors()) {
+      throw new InvalidParameterException(
+          Converter.convertFieldErrorListToJson(result.getFieldErrors()));
+    }
+    User user = userService.getUser(request.getId());
 
     return new UserResponse(user.getId(), user.getName(), user.getEmail(), user.isDelFlg());
   }
@@ -39,14 +49,19 @@ public class UserController {
   /**
    * Create user.
    *
-   * @param userRequest UserSearchRequest (name, email)
+   * @param request UserCreateRequest (name, email)
    * @return UserResponse (id, name, email, delFlg)
    */
-  @PostMapping(value = "/insert")
-  public UserResponse insert(@ModelAttribute UserRequest userRequest)
+  @PostMapping(value = "/create")
+  public UserResponse create(@ModelAttribute @Validated UserCreateRequest request,
+      BindingResult result)
       throws Exception {
+    if (result.hasErrors()) {
+      throw new InvalidParameterException(
+          Converter.convertFieldErrorListToJson(result.getFieldErrors()));
+    }
     User user =
-        new User(null, userRequest.getName(), userRequest.getEmail(), LocalDateTime.now(), null,
+        new User(null, request.getName(), request.getEmail(), LocalDateTime.now(), null,
             false);
     userService.createUser(user);
 
@@ -56,14 +71,19 @@ public class UserController {
   /**
    * Update user.
    *
-   * @param userRequest UserSearchRequest (id, name, email, delFlg)
+   * @param request UserUpdateRequest (id, name, email, delFlg)
    * @return UserResponse (id, name, email, delFlg)
    */
   @PostMapping(value = "/update")
-  public UserResponse update(@ModelAttribute UserRequest userRequest)
+  public UserResponse update(@ModelAttribute @Validated UserUpdateRequest request,
+      BindingResult result)
       throws Exception {
-    User user = userService.getUser(userRequest.getId());
-    userService.updateUser(user, userRequest);
+    if (result.hasErrors()) {
+      throw new InvalidParameterException(
+          Converter.convertFieldErrorListToJson(result.getFieldErrors()));
+    }
+    User user = userService.getUser(request.getId());
+    userService.updateUser(user, request);
 
     return new UserResponse(user.getId(), user.getName(), user.getEmail(), user.isDelFlg());
   }
@@ -71,13 +91,17 @@ public class UserController {
   /**
    * Delete user.
    *
-   * @param userRequest UserSearchRequest (id)
+   * @param request UserRequest (id)
    * @return UserResponse (id, name, email, delFlg)
    */
   @PostMapping(value = "/delete")
-  public UserResponse delete(@ModelAttribute UserRequest userRequest)
+  public UserResponse delete(@ModelAttribute @Validated UserRequest request, BindingResult result)
       throws Exception {
-    User user = userService.getUser(userRequest.getId());
+    if (result.hasErrors()) {
+      throw new InvalidParameterException(
+          Converter.convertFieldErrorListToJson(result.getFieldErrors()));
+    }
+    User user = userService.getUser(request.getId());
     userService.deleteUser(user.getId());
 
     return new UserResponse(user.getId(), user.getName(), user.getEmail(), true);
